@@ -9,7 +9,7 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "ant-design-vue": "^1.6.2",
+    "ant-design-vue": "^2.0.0-beta.15",
     "axios": "^0.19.2",
     "babel-runtime": "^6.26.0",
     "core-js": "^3.6.5",
@@ -22,18 +22,21 @@
     "three": "^0.110.0",
     "ts-loader": "^8.0.4",
     "typescript": "^4.0.3",
-    "vue": "^2.6.10",
+    "vue": "^3.0.0",
     "vue-draggable-resizable": "^2.1.0",
-    "vue-loader": "^15.7.1",
+    "vue-loader": "^16.0.0-rc.1",
     "vue-ls": "^3.2.1",
     "vue-markdown": "^2.2.4",
-    "vue-router": "^3.1.3",
-    "vuex": "^3.1.2"
+    "vue-router": "^4.0.0-rc.2",
+    "vuex": "^4.0.0-rc.1"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^4.4.6",
+    "@types/webpack-env": "^1.15.3",
+    "@vue/cli-plugin-babel": "^4.5.8",
+    "@vue/cli-plugin-router": "^4.5.8",
     "@vue/cli-plugin-unit-jest": "~4.5.0",
-    "@vue/cli-service": "^2.0.0",
+    "@vue/cli-service": "^4.0.0",
+    "@vue/compiler-sfc": "^3.0.2",
     "@vue/test-utils": "^1.0.3",
     "babel-plugin-import": "^1.13.0",
     "css-loader": "^3.6.0",
